server/requirements.txt
Django>=5.2.7
djangorestframework>=3.16.0
django-cors-headers>=4.9.0
mssql-django>=1.6.0
pyodbc>=5.3.0
